examples/lesson07-testing/projects/auth/src/app/home/home.component.ts
import { Component, OnInit } from '@angular/core';
import { FormBuilder, Validators } from '@angular/forms';

@Component({
  selector: 'app-home',
  templateUrl: './home.component.html',
  styleUrls: ['./home.component.scss']
})
export class HomeComponent {

  authForm = this.formBuilder.group({
    username: ['', [Validators.required]],
    password: ['', [Validators.required]],
  })

  constructor(private formBuilder: FormBuilder) { }

  onSubmit() {
    console.log(this.authForm.value)
  }

}
